+++ sdk/ai/azure-ai-resources/azure/ai/resources/_index/_dataindex/entities/_builders/data_index_func.py
@@ -224,14 +224,14 @@ def data_index_incremental_update_acs(
         input_data: Input,
         embeddings_model: str,
         acs_config: str,
-        acs_connection_id: str,
-        chunk_size: int = 768,
-        chunk_overlap: int = 0,
-        input_glob: str = "**/*",
-        citation_url: str = None,
-        citation_replacement_regex: str = None,
-        aoai_connection_id: str = None,
-        embeddings_container: Input = None,
+        acs_connection_id: Optional[str],
+        embeddings_container: Input,
+        chunk_size: Optional[int] = 768,
+        chunk_overlap: Optional[int] = 0,
+        input_glob: Optional[str] = "**/*",
+        citation_url: Optional[str] = None,
+        citation_replacement_regex: Optional[str] = None,
+        aoai_connection_id: Optional[str] = None,
     ):
         """
         Generate embeddings for a `input_data` source and push them into an Azure Cognitive Search index.
@@ -243,12 +243,12 @@ def data_index_incremental_update_acs(
         :param acs_config: The configuration for the Azure Cognitive Search index.
         :type acs_config: str
         :param acs_connection_id: The connection ID for the Azure Cognitive Search index.
-        :type acs_connection_id: str
-        :param chunk_size: The size of the chunks to break the input data into.
+        :type acs_connection_id: Optional[str]
+        :param chunk_size: The size of the chunks to break the input data into. Defaults to 768.
         :type chunk_size: int
-        :param chunk_overlap: The number of tokens to overlap between chunks.
+        :param chunk_overlap: The number of tokens to overlap between chunks. Defaults to 0.
         :type chunk_overlap: int
-        :param input_glob: The glob pattern to use when searching for input data.
+        :param input_glob: The glob pattern to use when searching for input data. Defaults to "**/*".
         :type input_glob: str
         :param citation_url: The URL to use when generating citations for the input data.
         :type citation_url: str
@@ -261,6 +261,13 @@ def data_index_incremental_update_acs(
         :return: The URI of the generated Azure Cognitive Search index.
         :rtype: str.
         """
+        if input_glob is None:
+            input_glob = "**/*"
+        if chunk_size is None:
+            chunk_size = 768
+        if chunk_overlap is None:
+            chunk_overlap = 0
+
         crack_and_chunk_and_embed = crack_and_chunk_and_embed_component(
             input_data=input_data,
             input_glob=input_glob,
@@ -359,7 +366,7 @@ def data_index_faiss(
     input_data_override: Optional[Input] = None,
 ):
     from azure.ai.resources._index._dataindex.data_index.models import build_model_protocol
-    from azure.ai.resources._indexx._dataindex.dsl._pipeline_decorator import pipeline
+    from azure.ai.resources._index._dataindex.dsl._pipeline_decorator import pipeline
 
     crack_and_chunk_component = get_component_obj(ml_client, LLMRAGComponentUri.LLM_RAG_CRACK_AND_CHUNK)
     generate_embeddings_component = get_component_obj(ml_client, LLMRAGComponentUri.LLM_RAG_GENERATE_EMBEDDINGS)
@@ -378,12 +385,12 @@ def data_index_faiss(
     def data_index_faiss_pipeline(
         input_data: Input,
         embeddings_model: str,
-        chunk_size: int = 1024,
-        data_source_glob: str = None,
-        data_source_url: str = None,
-        document_path_replacement_regex: str = None,
-        aoai_connection_id: str = None,
-        embeddings_container: Input = None,
+        embeddings_container: Input,
+        chunk_size: Optional[int] = 1024,
+        data_source_glob: Optional[str] = None,
+        data_source_url: Optional[str] = None,
+        document_path_replacement_regex: Optional[str] = None,
+        aoai_connection_id: Optional[str] = None,
     ):
         """
         Generate embeddings for a `input_data` source and create a Faiss index from them.
@@ -392,7 +399,7 @@ def data_index_faiss(
         :type input_data: Input
         :param embeddings_model: The embedding model to use when processing source data chunks.
         :type embeddings_model: str
-        :param chunk_size: The size of the chunks to break the input data into.
+        :param chunk_size: The size of the chunks to break the input data into. Defaults to 1024.
         :type chunk_size: int
         :param data_source_glob: The glob pattern to use when searching for input data.
         :type data_source_glob: str
@@ -407,6 +414,9 @@ def data_index_faiss(
         :return: The URI of the generated Faiss index.
         :rtype: str.
         """
+        if chunk_size is None:
+            chunk_size = 1024
+
         crack_and_chunk = crack_and_chunk_component(
             input_data=input_data,
             input_glob=data_source_glob,
@@ -517,14 +527,14 @@ def data_index_acs(
         input_data: Input,
         embeddings_model: str,
         acs_config: str,
-        acs_connection_id: str,
-        chunk_size: int = 1024,
-        data_source_glob: str = None,
-        data_source_url: str = None,
-        document_path_replacement_regex: str = None,
-        aoai_connection_id: str = None,
-        embeddings_container: Input = None,
-    ):
+        acs_connection_id: Optional[str],
+        embeddings_container: Input,
+        chunk_size: Optional[int] = 1024,
+        data_source_glob: Optional[str] = None,
+        data_source_url: Optional[str] = None,
+        document_path_replacement_regex: Optional[str] = None,
+        aoai_connection_id: Optional[str] = None,
+    ) -> Dict[str, Any]:
         """
         Generate embeddings for a `input_data` source and push them into an Azure Cognitive Search index.
 
@@ -536,7 +546,7 @@ def data_index_acs(
         :type acs_config: str
         :param acs_connection_id: The connection ID for the Azure Cognitive Search index.
         :type acs_connection_id: str
-        :param chunk_size: The size of the chunks to break the input data into.
+        :param chunk_size: The size of the chunks to break the input data into. Defaults to 1024.
         :type chunk_size: int
         :param data_source_glob: The glob pattern to use when searching for input data.
         :type data_source_glob: str
@@ -549,8 +559,11 @@ def data_index_acs(
         :param embeddings_container: The container to use when caching embeddings.
         :type embeddings_container: Input
         :return: The URI of the generated Azure Cognitive Search index.
-        :rtype: str.
+        :rtype: Dict[str, Any]
         """
+        if chunk_size is None:
+            chunk_size = 1024
+
         crack_and_chunk = crack_and_chunk_component(
             input_data=input_data,
             input_glob=data_source_glob,